requirements.txt
Django==2.0.4
psycopg2
djangorestframework
markdown
django-filter
django-debug-toolbar
requests
pyjwt
pycrypto
cryptography
python-jose